requirements.txt
flask
flask-socketio
eventlet
gunicorn
pycryptodome
psycopg2-binary
